Add method to deal with empty or duplicate constraint names
Add method to cons_decomp that is called in SCIP_STAGE_PROBLEM and
repairs empty and duplicate constraint names. Repairing is done by
giving a generic name that is not used as constraint name yet of
the form "c_"+n such that n is minimal natural number.
This method is called in main.c if gcg is called via command line
with -f option.